requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
line-bot-sdk==2.4.1
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
